Restructure web folders
