+++ comscore/src/internal/web/SkeletonApi.ts
