--- comscore/src/internal/web/index.ts
@@ -1,15 +1,0 @@
-import { ComscoreTheo } from './integration/ComscoreTheo';
-import { ContentMetadata } from './comscore/ContentMetadata';
-import { AdMetadata } from './comscore/AdMetadata';
-import {
-  ComscoreSkeletonMock,
-  ComscoreSkeletonWeb,
-} from './comscore/SkeletonApi';
-
-(window as any).THEOplayerComscoreSDK = {
-  ComscoreTheo,
-  ContentMetadata,
-  AdMetadata,
-  ComscoreSkeletonMock,
-  ComscoreSkeletonWeb,
-};
